esun_api_docker/app/requirements.txt
flask
numpy
opencv-python
albumentations
Pillow
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.1+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.8.1